% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcf.R
\name{mcf}
\alias{mcf}
\title{Calculates McFadden's Pseudo R-Squared}
\usage{
mcf(model)
}
\arguments{
\item{model}{a logit model from which you would like to return McFadden's pseudo r-squared. This can be a model created either with 'glm.mids()' or 'with()'}
}
\value{
mcfs2: McFadden's pseudo r-squared
}
\description{
Returns McFadden's pseudo r-squared for logistic regression models performed on 'mice'-imputed data sets.
}
