% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modds.R
\name{modds}
\alias{modds}
\title{Calculates an OR and associated probability for an 'average' case}
\usage{
modds(x)
}
\arguments{
\item{x}{a logit model performed using 'glm(...,family=binomial(link="logit"))'}
}
\value{
odds: an odds ratio; prob: the associated probability analogous to this odds ratio
}
\description{
Returns the odds ratio and associated probability for a user-defined outcome (using logistic regression; 'glm()') for a case whose values are all at their median, estimating an 'average' case.
}
