% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preds.R
\name{preds}
\alias{preds}
\title{Prediction Accuracy and Diagnostic Statistics for Logistic Regression Models}
\usage{
preds(x, plevel = 0.5)
}
\arguments{
\item{x}{a logistic regression model performed using 'glm(...family=binomial(link="logit"))'}

\item{plevel}{the level at which you would like to classify cases vs. non-cases (default is p=0.5)}
}
\value{
accuracy of predictions, chi-square test of model vs. null, sensitivity, specificity, % advantage in classifications using model vs. null model.
}
\description{
Returns accuracy of predictions, a chi-square test of model vs. null model, sensitivity, specificity, and % accuracy of predictions vs. null model for logistic regression models.
}
