% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preds_m.R
\name{preds_m}
\alias{preds_m}
\title{Prediction Accuracy and Diagnostic Statistics for Logistic Regression Models Performed on 'mice'-Imputed Data Sets}
\usage{
preds_m(model, imp, plevel = 0.5)
}
\arguments{
\item{model}{a logistic regression model performed using either 'glm.mids()' or 'with()', from package 'mice'}

\item{imp}{the object to which your imputed data sets are assigned (i.e., the object that took the output of the mice() function)}

\item{plevel}{the level at which you would like to classify cases vs. non-cases (default is p=0.5)}
}
\value{
accuracy of predictions, chi-square test of model vs. null, sensitivity, specificity, % advantage in classifications using model vs. null model.
}
\description{
Returns accuracy of predictions, a chi-square test of model vs. null model, sensitivity, specificity, and % accuracy of predictions vs. null model for logistic regression models performed on 'mice'-imputed data sets. This will use pooled coefficients from all models to calculate the accuracy of correct classifications using your model.
}
