% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity Estimate for Logistic Regression Models.}
\usage{
sensitivity(predicted, actual)
}
\arguments{
\item{predicted}{this parameter can either take (i) a 2x2 table/confusion matrix of predictions and observed classes (with predictions displayed on the vertical, left side), or (ii) your predicted classifications.}

\item{actual}{if param predicted is not a table, this parameter will take your actual (observed) values of your outcome (must be binary, not the predicted probabilities)}
}
\value{
sensitivity estimate
}
\description{
Returns a sensitvity estimate when comparing your predicted classifications to observed classifications.
}
