% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{specificity}
\alias{specificity}
\title{Specificity Estimate for Logistic Regression Models.}
\usage{
specificity(predicted, actual)
}
\arguments{
\item{predicted}{this parameter can either take (i) a 2x2 table/confusion matrix of predictions and observed classes (with predictions displayed on the vertical, left side), or (ii) your predicted classifications.}

\item{actual}{if param predicted is not a table, this parameter will take your actual (observed) values of your outcome (must be binary, not the predicted probabilities)}
}
\value{
specificity estimate
}
\description{
Returns a specificity estimate when comparing your predicted classifications to observed classifications.
}
